fixed id types and created first controller

+++ ZooApp.Application/Commands/CreateAnimalCommand.cs
@@ -8,9 +8,9 @@ namespace ZooApp.Application.Commands;
 public record CreateAnimalCommand
 {
     public string Name { get; init; }
-    public ulong SpeciesId { get; init; }
+    public int SpeciesId { get; init; }
 
-    public CreateAnimalCommand(string name, ulong speciesId)
+    public CreateAnimalCommand(string name, int speciesId)
     {
         Name = name;
         SpeciesId = speciesId;
