+++ ZooApp.Domain/ZooKeeper/IZooKeeperRepository.cs
@@ -7,10 +7,10 @@ namespace ZooApp.Domain.ZooKeeper;
 
 public interface IZooKeeperRepository
 {
-    Task<ZooKeeper?> GetById(ulong id);
-    Task<ulong> DispatchTaskAutomatically(DateTime scheduledAt, TimeSpan duration, string description); //tu będzie funkcja/procedura składowana do bazy danych
+    Task<ZooKeeper?> GetById(int id);
+    Task<int> DispatchTaskAutomatically(DateTime scheduledAt, TimeSpan duration, string description); //tu będzie funkcja/procedura składowana do bazy danych
     Task<ZooKeeper> Save(ZooKeeper zooKeeper);
     Task Delete(ZooKeeper zooKeeper);
-    Task<List<ITask>> GetTasksForZooKeeper(ulong zooKeeperId);
-    Task<List<ITask>> GetTasksForZooKeeperForTheDate(ulong zooKeeperId, DateTime date);
+    Task<List<ITask>> GetTasksForZooKeeper(int zooKeeperId);
+    Task<List<ITask>> GetTasksForZooKeeperForTheDate(int zooKeeperId, DateTime date);
 }
