+++ ZooApp.Domain/ZooKeeper/Tasks/ITask.cs
@@ -6,8 +6,8 @@ namespace ZooApp.Domain.ZooKeeper.Tasks;
 
 public interface ITask
 {
-    ulong Id { get; init; }
-    ulong ZooKeeperId { get; } // some people say that id breaks the DDD principles, some say it doesn't, but surely it makes mapping to DB easier
+    int Id { get; init; }
+    int ZooKeeperId { get; } // some people say that id breaks the DDD principles, some say it doesn't, but surely it makes mapping to DB easier
     DateTime ScheduledAt { get;}
     TimeSpan Duration { get;}
     string Description { get;}
